+++ src/renderer/components/uielements/assets/assetCard/AssetCard.tsx
@@ -5,8 +5,8 @@ import { Dropdown } from 'antd'
 import BigNumber from 'bignumber.js'
 import { sortBy as _sortBy } from 'lodash'
 
+import { ZERO_BN } from '../../../../const'
 import { useClickOutside } from '../../../../hooks/useOutsideClick'
-import { ZERO_BN } from '../../../../services/const'
 import { PriceDataIndex } from '../../../../services/midgard/types'
 import { AssetPair } from '../../../../types/asgardex'
 import Label from '../../label'
