+++ src/renderer/components/uielements/input/InputBigNumber.tsx
@@ -5,8 +5,8 @@ import BigNumber from 'bignumber.js'
 import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
 
+import { ZERO_BN } from '../../../const'
 import { trimZeros } from '../../../helpers/stringHelper'
-import { ZERO_BN } from '../../../services/const'
 import * as Styled from './Input.style'
 import { VALUE_ZERO, formatValue, validInputValue } from './util'
