+++ src/renderer/helpers/assetHelper.test.ts
@@ -1,6 +1,8 @@
 import { assetFromString, EMPTY_ASSET } from '@thorchain/asgardex-util'
 
-import { isRuneAsset, isBnbAsset } from './assetHelper'
+import { CURRENCY_SYMBOLS } from '../const'
+import { PoolAsset } from '../views/pools/types'
+import { isRuneAsset, isBnbAsset, getCurrencySymbolByAssetString } from './assetHelper'
 
 describe('helpers/assetHelper', () => {
   describe('isRuneAsset', () => {
@@ -28,6 +30,20 @@ describe('helpers/assetHelper', () => {
     it('returns false for any other asset than BNB', () => {
       const asset = assetFromString('BNB.RUNE-B1A') || EMPTY_ASSET
       expect(isBnbAsset(asset)).toBeFalsy()
+    })
+  })
+
+  describe('getCurrencySymbolByAssetString', () => {
+    it('should return rune symbol', () => {
+      expect(getCurrencySymbolByAssetString(PoolAsset.RUNEB1A)).toEqual(CURRENCY_SYMBOLS['BNB.RUNE-67C'])
+    })
+
+    it('should return ticker for unknown asset', () => {
+      expect(getCurrencySymbolByAssetString('chain.ticker-symbol')).toEqual('ticker')
+    })
+
+    it('should return empty string for invalid asset', () => {
+      expect(getCurrencySymbolByAssetString('invalid asset')).toEqual('')
     })
   })
 })
