+++ src/renderer/services/midgard/pools.ts
@@ -1,5 +1,6 @@
 import * as RD from '@devexperts/remote-data-ts'
 import { Asset, assetToString, bn } from '@thorchain/asgardex-util'
+import BigNumber from 'bignumber.js'
 import * as A from 'fp-ts/Array'
 import * as FP from 'fp-ts/function'
 import { some } from 'fp-ts/Option'
@@ -9,14 +10,15 @@ import * as Rx from 'rxjs'
 import { combineLatest } from 'rxjs'
 import { catchError, map, retry, shareReplay, startWith, switchMap, filter } from 'rxjs/operators'
 
-import { PRICE_POOLS_WHITELIST } from '../../const'
+import { ONE_BN, PRICE_POOLS_WHITELIST } from '../../const'
 import { sequenceTOption } from '../../helpers/fpHelpers'
 import { LiveData, liveData } from '../../helpers/rx/liveData'
 import { observableState, triggerStream } from '../../helpers/stateHelper'
 import { DefaultApi, GetPoolsDetailsViewEnum } from '../../types/generated/midgard/apis'
+import { PoolDetail } from '../../types/generated/midgard/models'
 import { isPricePoolAsset, PricePoolAsset, RUNEAsset } from '../../views/pools/types'
 import { getCurrentNetworkState, network$ } from '../app/service'
-import { mapNetworkToPoolAssets, MIDGARD_MAX_RETRY, ONE_BN } from '../const'
+import { mapNetworkToPoolAssets, MIDGARD_MAX_RETRY } from '../const'
 import { PoolsStateRD, SelectedPricePoolAsset } from './types'
 import { getPricePools, pricePoolSelector } from './utils'
 
@@ -70,7 +72,7 @@ const createPoolsService = (
   /**
    * Get `PoolDetails` data from Midgard
    */
-  const apiGetPoolsData$ = (asset: string, isDetailed = false) =>
+  const apiGetPoolsData$ = (asset: string, isDetailed = false): LiveData<Error, PoolDetail[]> =>
     byzantine$.pipe(
       liveData.chain((endpoint) =>
         pipe(
@@ -151,7 +153,7 @@ const createPoolsService = (
   // `TriggerStream` to reload detailed data of pool
   const { get$: reloadPoolDetailedState$, set: reloadPoolDetailedState } = observableState<O.Option<Asset>>(O.none)
 
-  const poolDetailedState$ = reloadPoolDetailedState$.pipe(
+  const poolDetailedState$: LiveData<Error, PoolDetail> = reloadPoolDetailedState$.pipe(
     filter(O.isSome),
     switchMap((asset) => apiGetPoolsData$(assetToString(asset.value), true)),
     liveData.chain(
@@ -193,9 +195,9 @@ const createPoolsService = (
   /**
    * Use this to convert asset's price to selected price asset by multiplying to the priceRation inner value
    */
-  const priceRatio$ = pipe(
+  const priceRatio$: Rx.Observable<BigNumber> = pipe(
     combineLatest([pipe(poolsState$, map(RD.toOption)), selectedPricePoolAsset$]),
-    map((state) => sequenceTOption(state[0], state[1])),
+    map(([pools, selectedAsset]) => sequenceTOption(pools, selectedAsset)),
     map(
       O.chain(([pools, selectedAsset]) =>
         pipe(
@@ -204,9 +206,8 @@ const createPoolsService = (
         )
       )
     ),
-    map(O.map((detail) => bn(detail.priceRune || 1))),
-    map(O.getOrElse(() => ONE_BN)),
-    map((price) => ONE_BN.dividedBy(price))
+    map(O.map((detail) => ONE_BN.dividedBy(bn(detail.priceRune || 1)))),
+    map(O.getOrElse(() => ONE_BN))
   )
 
   return {
